UploadCameraImageEdgeToCloud/UploadCameraImageEdgeToCloud/modules/StoreImageToLocalBlob/requirements.txt
azure-iothub-device-client~=1.4.3
flask
azure-storage-blob
azure-mgmt-storage
